moved ShowConfirmationDialog and  ShowFolderSelectionDialog to separate files
