+++ app/src/main/java/com/olup/notable/components/ConfirmationDialog.kt
@@ -1,0 +1,51 @@
+package com.olup.notable.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.RectangleShape
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.window.Dialog
+import com.olup.notable.modals.ActionButton
+
+
+@Composable
+fun ShowConfirmationDialog(
+    title: String,
+    message: String,
+    onConfirm: () -> Unit,
+    onCancel: () -> Unit
+) {
+    Dialog(onDismissRequest = { onCancel() }) {
+        Column(
+            modifier = Modifier
+                .background(Color.White)
+                .border(1.dp, Color.Black, RectangleShape),
+            verticalArrangement = Arrangement.spacedBy(12.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text(text = title, fontWeight = FontWeight.Bold, fontSize = 20.sp)
+            Text(text = message, fontSize = 16.sp)
+            Row(
+                horizontalArrangement = Arrangement.SpaceAround,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(24.dp)
+            ) {
+                ActionButton(text = "Cancel", onClick = onCancel)
+                ActionButton(text = "Confirm", onClick = onConfirm)
+            }
+        }
+    }
+}
